+++ app/src/main/java/camp/nextstep/DispatcherServlet.java
@@ -1,5 +1,6 @@
 package camp.nextstep;
 
+import camp.nextstep.manual.ManualHandlerMapping;
 import com.interface21.webmvc.servlet.ModelAndView;
 import com.interface21.webmvc.servlet.mvc.asis.Controller;
 import com.interface21.webmvc.servlet.mvc.tobe.annotation.AnnotationHandlerMapping;
